+++ packages/google_maps_flutter/google_maps_flutter/lib/src/google_map.dart
@@ -123,6 +123,7 @@ class GoogleMap extends StatefulWidget {
     this.onCameraIdle,
     this.onTap,
     this.onLongPress,
+    this.onPoiTap,
   })  : assert(initialCameraPosition != null),
         super(key: key);
 
@@ -224,6 +225,9 @@ class GoogleMap extends StatefulWidget {
 
   /// Called every time a [GoogleMap] is long pressed.
   final ArgumentCallback<LatLng>? onLongPress;
+
+  /// Called every time a [PointOfInterest] is clicked.
+  final ArgumentCallback<PointOfInterest>? onPoiTap;
 
   /// True if a "My Location" layer should be shown on the map.
   ///
@@ -527,6 +531,14 @@ class _GoogleMapState extends State<GoogleMap> {
     final ArgumentCallback<LatLng>? onLongPress = widget.onLongPress;
     if (onLongPress != null) {
       onLongPress(position);
+    }
+  }
+
+  void onPoiTap(PointOfInterest poi) {
+    assert(poi != null);
+    final ArgumentCallback<PointOfInterest>? onPoiTap = widget.onPoiTap;
+    if (onPoiTap != null) {
+      onPoiTap(poi);
     }
   }
 }